APPS CON ESTRUCTURA/FORENSIKvIDEODETECTOR2/requirements.txt
streamlit
opencv-python
opencv-contrib-python
imagehash
numpy
Pillow
reportlab
matplotlib
tqdm
scikit-learn
scikit-image
pandas
plotly
python-docx
weasyprint
jinja2
markdown2
bleach
pathlib2
tensorflow
torch
ffmpeg-python
networkx
seaborn